djangoapp/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more
# information.

django-appconf
django-extensions
django-cachalot

rjsmin
rcssmin
django-bower
django-pipeline

django-robots

django-storages>1.5
boto

django-toolbelt
whitenoise
waitress
